refactor!: more refactoring (squash this commit)

+++ alchemist-loading/src/test/kotlin/it/unibo/alchemist/test/TestGraphStreamReproducibility.kt
@@ -13,9 +13,9 @@ import io.kotest.core.spec.style.FreeSpec
 import io.kotest.matchers.shouldBe
 import it.unibo.alchemist.loader.GraphStreamSupport
 import it.unibo.alchemist.model.SupportedIncarnations
-import it.unibo.alchemist.model.environments.Continuous2DEnvironment
+import it.unibo.alchemist.model.euclidean.positions.Euclidean2DPosition
 import it.unibo.alchemist.model.nodes.GenericNode
-import it.unibo.alchemist.model.positions.Euclidean2DPosition
+import it.unibo.alchemist.model.physics.environments.Continuous2DEnvironment
 import org.apache.commons.math3.random.MersenneTwister
 
 typealias EnvironmentDisplacement = List<Pair<List<Double>, List<Int>>>
@@ -40,7 +40,7 @@ class TestGraphStreamReproducibility : FreeSpec({
                         environment = environment,
                         nodeCount = 100,
                         generatorName = graphType,
-                        uniqueId = uniqueId.toLong(),
+                        uniqueId = uniqueId,
                         layoutQuality = 0.1,
                         parameters = parameters.toTypedArray(),
                     )
